requirements.txt
# Core dependencies
fastapi==0.115.12
httpx==0.28.1
uvicorn==0.41.0
pydantic==2.12.5
python-multipart==0.0.22

# OCR and image processing
paddleocr==3.4.0
paddlepaddle==3.3.0
manga-ocr==0.1.14
pillow==12.1.1
opencv-contrib-python==4.10.0.84

# Machine learning
torch==2.10.0
transformers==5.3.0
numpy==2.2.6

# Utilities
loguru==0.7.3
requests==2.32.5
tqdm==4.67.3
pyyaml==6.0.2
regex==2026.2.28
shapely==2.1.2
pyclipper==1.4.0

# Additional dependencies
modelscope==1.34.0
paddlex==3.4.2
pandas==2.3.3
networkx==3.4.2
protobuf==7.34.0
safetensors==0.7.0
tokenizers==0.22.2
rich==14.3.3
psutil==7.2.2
py-cpuinfo==9.0.0
prettytable==3.17.0
ujson==5.11.0
pyperclip==1.11.0
pypdfium2==5.5.0
python-bidi==0.6.7
unidic-lite==1.0.8
pycryptodome==3.23.0
opt-einsum==3.3.0
mpmath==1.3.0
sympy==1.14.0
typing-extensions==4.15.0
typing-inspection==0.4.2
shellingham==1.5.4
typer==0.24.1
termcolor==3.3.0
wcwidth==0.6.0
urllib3==2.6.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
tzdata==2025.3
ruamel-yaml==0.19.1
markdown-it-py==4.0.0
mdurl==0.1.2
markupsafe==3.0.3
pygments==2.19.2
six==1.17.0
win32-setctime==1.2.0
packaging==25.0
